move KeypadKey component to its own file
